+++ openeo_driver/backend.py
@@ -388,7 +388,7 @@ class BatchJobs(MicroService):
     ASSET_PUBLIC_HREF = "public_href"
 
     def __init__(self):
-        # TODO this "proxy user" feature is YARN/Spark/VITO specific. Move it to oppeno-geopyspark-driver?
+        # TODO #275 this "proxy_user" feature is YARN (VITO) specific. Move it to oppeno-geopyspark-driver?
         self._get_proxy_user: Callable[[User], Optional[str]] = lambda user: None
 
     def create_job(
@@ -503,11 +503,12 @@ class BatchJobs(MicroService):
         """
         raise NotImplementedError
 
-    # TODO this "proxy user" feature is YARN/Spark/VITO specific. Move it to oppeno-geopyspark-driver?
     def get_proxy_user(self, user: User) -> Optional[str]:
+        # TODO #275 this "proxy_user" feature is YARN (VITO) specific. Move it to oppeno-geopyspark-driver?
         return self._get_proxy_user(user)
 
     def set_proxy_user_getter(self, getter: Callable[[User], Optional[str]]):
+        # TODO #275 this "proxy_user" feature is YARN (VITO) specific. Move it to oppeno-geopyspark-driver?
         self._get_proxy_user = getter
 
 
@@ -750,8 +751,13 @@ class OpenEoBackendImplementation:
     def summarize_exception(self, error: Exception) -> Union[ErrorSummary, Exception]:
         return error
 
-    # TODO this "proxy user" feature is YARN/Spark/VITO specific. Move it to oppeno-geopyspark-driver?
     def set_preferred_username_getter(self, getter: Callable[[User], Optional[str]]):
+        # TODO #275 this "proxy_user" feature is YARN/Spark/VITO specific. Move it to oppeno-geopyspark-driver?
+        # TODO this method talks about "preferred_username" but is currently used to propagate
+        #      the YARN proxy_user (which must be unique and does align very well with a user chosen preferred username)
+        #      In other cases (eu-cdse/openeo-cdse-infra#56) we might generate the YARN proxy_user
+        #      in an even more constrained way, completely separate from any preferred username.
+        #      In short: this method's name might set wrong expectations.
         self.batch_jobs.set_proxy_user_getter(getter)
 
     def user_access_validation(self, user: User, request: flask.Request) -> User: