+++ openeo_driver/delayed_vector.py
@@ -10,6 +10,7 @@ import fiona
 import geopandas as gpd
 import pyproj
 import requests
+import shapely
 from shapely.geometry import shape
 from shapely.geometry.base import BaseGeometry
 
@@ -121,6 +122,45 @@ class DelayedVector:
                 for geometry in geometries
             ]
         return geometries
+
+    @property
+    def geojson(self) -> dict:
+        # TODO: Also reproject geojson files to wgs84 if required.
+        if self.path.startswith("http"):
+            if not DelayedVector._is_shapefile(self.path):
+                geojson = self._load_geojson_url(url=self.path)
+                return self._geojson_to_wgs84(geojson)
+        else:  # it's a file on disk
+            if not self.path.endswith(".shp"):
+                with open(self.path, "r") as f:
+                    geojson = json.load(f)
+                    return self._geojson_to_wgs84(geojson)
+        geometries = self.geometries_wgs84
+        return {
+            "type": "FeatureCollection",
+            "features": [
+                {"type": "Feature", "geometry": shapely.geometry.mapping(geometry), "properties": {}}
+                for geometry in geometries
+            ],
+        }
+
+    def _geojson_to_wgs84(self, geojson: dict) -> dict:
+        wgs84 = pyproj.CRS("EPSG:4326")
+        if self.crs != wgs84:
+            # Reproject the geometries in the geojson but keep the properties.
+            geometries = [
+                reproject_geometry(shape(feature["geometry"]), from_crs=self.crs, to_crs=wgs84)
+                for feature in geojson["features"]
+            ]
+            features = []
+            for geometry, feature in zip(geometries, geojson["features"]):
+                feature["geometry"] = shapely.geometry.mapping(geometry)
+                feature_dict = {"type": "Feature", "geometry": feature["geometry"]}
+                if "properties" in feature:
+                    feature_dict["properties"] = feature["properties"]
+                features.append(feature_dict)
+            return {"type": "FeatureCollection", "features": features}
+        return geojson
 
     @property
     def area(self):